Moved 'wrong msg in wrong state' test
