+++ src/mgr/hermes_test_manager/hermes_test_api.py
@@ -15,7 +15,6 @@ from test_cases import EnvironmentManager
 # pylint: disable=unused-import
 import test_cases.test_cases_dummy
 import test_cases.test_downstream_ifc
-import test_cases.test_downstream_ifc_states
 import test_cases.test_downstream_ifc_interactive
 import test_cases.test_upstream_ifc
